requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
trimesh==4.0.5
numpy==1.24.3
pydantic==2.5.0
numpy-stl==3.1.0
requests==2.31.0 